+++ CHANGELOG.md
@@ -1,9 +1,17 @@
+# 1.1.3
+
+- Better error messages
+- New error action notes
+- Revert behavior of auto-creating new `.scss` element when one not found
+
 # 1.1.2
+
 - UI style updates to better match the new Webflow design
 - Fix editor height whenever iframe window resizes
 - Tab indentation in the editor
 
 # 1.1.1
+
 - Finally fix bundle 404 error - add relative path to build config
 
 # 1.1.0